+++ src/main/resources/db/migration/V3__split_purchase_order_into_lines.sql
@@ -7,7 +7,7 @@ CREATE TABLE IF NOT EXISTS purchase_order_lines (
     id INTEGER NOT NULL DEFAULT nextval('purchase_order_line_sequence') PRIMARY KEY,
     purchase_order_id INTEGER,
     product_id INTEGER,
-    quantity INTEGER,
+    quantity DECIMAL(10, 2),
     unit_price DOUBLE PRECISION NOT NULL,
     CONSTRAINT fk_product_3 FOREIGN KEY (product_id) REFERENCES Products(id),
     CONSTRAINT fk_purchase_order FOREIGN KEY (purchase_order_id) REFERENCES purchase_orders(id)
